# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/forums": {
            "get": {
                "description": "GetAllPublicFroums",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forums"
                ],
                "summary": "GetAllPublicFroums",
                "responses": {
                    "200": {
                        "description": "获取全部公开论坛",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.StatusOKResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/controllers.ForumResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusInternalServerError"
                        }
                    }
                }
            },
            "post": {
                "description": "CreateForum",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forums"
                ],
                "summary": "CreateForum",
                "parameters": [
                    {
                        "description": "论坛名",
                        "name": "forum_name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "是否公开",
                        "name": "is_public",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "description": "论坛描述",
                        "name": "description",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "论坛创建成功",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusOKResponse"
                        }
                    },
                    "400": {
                        "description": "参数不合法",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusBadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusInternalServerError"
                        }
                    }
                }
            }
        },
        "/forums/{forum_id}": {
            "get": {
                "description": "GetForumByID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forums"
                ],
                "summary": "GetForumByID",
                "responses": {
                    "200": {
                        "description": "获取全部公开论坛",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.StatusOKResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/controllers.ForumResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusInternalServerError"
                        }
                    }
                }
            }
        },
...
        "/users": {
            "get": {
                "description": "GetAllUsers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "GetAllUsers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "将token放在请求头部的‘Authorization‘字段中，并以‘Bearer ‘开头",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户名的子串",
                        "name": "username",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取全部用户",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.StatusOKResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.User"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusInternalServerError"
                        }
                    }
                }
            },
            "put": {
                "description": "UserLogin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "UserLogin",
                "parameters": [
                    {
                        "description": "用户名或者邮箱",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "密码",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "正确登陆",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.StatusOKResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/controllers.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数不合法",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusBadRequestResponse"
                        }
                    },
                    "403": {
                        "description": "该用户名或邮箱不存在",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusForbiddenResponse"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusInternalServerError"
                        }
                    }
                }
            },
            "post": {
                "description": "UserRegister",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "UserRegister",
                "parameters": [
                    {
                        "description": "用户名",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "密码",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "邮箱",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注册成功",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusOKResponse"
                        }
                    },
                    "400": {
                        "description": "参数不合法",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusBadRequestResponse"
                        }
                    },
                    "403": {
                        "description": "该邮箱已经被使用",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusForbiddenResponse"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusInternalServerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.ForumResponse": {
            "type": "object",
            "properties": {
                "forums": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Forum"
                    }
                },
                "user_detail": {
                    "$ref": "#/definitions/models.UserDetail"
                }
            }
        },
        "controllers.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "models.Forum": {
            "type": "object",
            "properties": {
                "admin_list": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "cover": {
                    "type": "string"
                },
                "create_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "forum_id": {
                    "type": "integer"
                },
                "forum_name": {
                    "type": "string"
                },
                "is_public": {
                    "type": "boolean"
                },
                "post_num": {
                    "type": "integer"
                },
                "subscribe_num": {
                    "type": "integer"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "create_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "is_admin": {
                    "type": "boolean"
                },
                "password": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.UserDetail": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "create_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "is_admin": {
                    "type": "boolean"
                },
                "participate_list": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "password": {
                    "type": "string"
                },
                "star_list": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "responses.StatusBadRequestResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "responses.StatusForbiddenResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 403
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "responses.StatusInternalServerError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "responses.StatusOKResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}